# apps/jakarta-helidon-virtual-threads-demo/helidon-credit-service/src/main/java/com/example/creditdemo/helidon/CreditDecisionRequest.java
package com.example.creditdemo.helidon;

public record CreditDecisionRequest(
        String customerId,
        double requestedAmount,
        int termMonths,
        double annualIncome,
        double monthlyDebt,
        int creditScore
) {
}